requirements-dev.txt
flake8==3.8.4
ipython==8.1.1

mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0

# Deps do pytest
attrs==20.3.0
py==1.10.0
pluggy==0.13.1
pytest-mock==3.3.1

#Coverage
coverage==5.3
codecov==2.1.11
pytest-cov==2.10.1
pytest-django==4.1.0


-r requirements.txt
